rumqttc: move old code to v4 module
