+++ benchmarks/clients/rumqttsync.rs
@@ -1,4 +1,4 @@
-use rumqttc::{self, Client, Event, Incoming, MqttOptions, QoS};
+use rumqttc::v4::{Client, Event, Incoming, MqttOptions, QoS};
 use std::error::Error;
 use std::thread;
 use std::time::{Duration, Instant};
