+++ rumqttc/src/v4/eventloop.rs
@@ -1,6 +1,6 @@
-use crate::{framed::Network, Transport};
-use crate::{tls, Incoming, MqttState, Packet, Request, StateError};
-use crate::{MqttOptions, Outgoing};
+use crate::v4::{framed::Network, Transport};
+use crate::v4::{tls, Incoming, MqttState, Packet, Request, StateError};
+use crate::v4::{MqttOptions, Outgoing};
 
 use async_channel::{bounded, Receiver, Sender};
 #[cfg(feature = "websocket")]
